# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Simple POS API
  version: 1.0.0
  description: |-
    Contract-first API for Simple POS (ADR-0008). Precedence: BUSINESS_RULES.md (BR v1.2) > this file > the UI/UX PDF > code.

    **Hosts**: API `https://api.cashier.hossam-ameen.online`, PWA `https://cashier.hossam-ameen.online` (separate origins,
    CORS limited to the PWA origin, ADR-0003).

    **Conventions** (ADR-0005, 0008, 0010, 0013)
    - Paths have no trailing slash. JSON is snake_case. Ids are integers. Timestamps are RFC 3339 UTC.
    - **Money is always an integer number of minor units (piastres) in a field ending in `_minor`** (BR-GEN-01). The server
      computes every total, change, expected cash and difference (BR-GEN-03).
    - Success body: `{success: true, message, data}`; lists: `data = {items, pagination, …}`; 204 has no body.
    - Error body: `{success: false, code, message, details}`; `code` is a BR §9 code with its BR §9 HTTP status and Arabic
      `message` (BR-GEN-06).
    - Auth: `Authorization: Bearer <access_token>` (15 min). The refresh token is an httpOnly cookie on the API host only,
      used by `/api/auth/refresh` and `/api/auth/logout` with `X-Requested-With: simple-pos` (ADR-0003, ADR-0011).
    - `x-roles` lists the roles allowed to call each operation (`PUBLIC`, `ANY_AUTHENTICATED`, `ADMIN`, `CASHIER`); every
      other authenticated role gets `FORBIDDEN_ROLE` (BR-ROLE-05). `x-br` lists the rules the operation enforces and
...

...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /api/tables:
    get:
      operationId: listTables
      tags:
      - tables
      summary: Tables grid (screen 06)
      description: |-
        Ordered by number. `status` is derived: OCCUPIED iff an OPEN order references the table (BR-TBL-01); an occupied table
        carries its order number, item count (Σ qty) and total after discount (BR-TBL-04, OQ-14). CASHIER: active tables only
        (BR-TBL-06); ADMIN: all tables, `is_active` filter allowed (US-21.5). Counts respect the caller's visibility.

        BR: BR-TBL-01, BR-TBL-04, BR-TBL-06, OQ-14, AC-06.
        Errors: `UNAUTHENTICATED`.
      x-roles:
      - ADMIN
      - CASHIER
      x-br:
      - BR-TBL-01
      - BR-TBL-04
      - BR-TBL-06
      - OQ-14
      - AC-06
      x-error-codes:
      - UNAUTHENTICATED
      parameters:
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/TableStatus'
      - name: is_active
        in: query
        required: false
        schema:
          type: boolean
        description: ADMIN only; ignored for CASHIER.
      responses:
        '200':
          description: Tables.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableListEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
    post:
      operationId: createTable
      tags:
      - tables
      summary: Add a table (screen 06 admin dialog)
      description: |-
        Positive unique number (`DUPLICATE_VALUE`); created active by default (US-21.1, OQ-32). ADMIN only.

        BR: BR-ROLE-01, BR-TBL-06, OQ-32.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `DUPLICATE_VALUE`, `VALIDATION_ERROR`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      - BR-TBL-06
      - OQ-32
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - DUPLICATE_VALUE
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableCreate'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
  /api/tables/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      operationId: getTable
      tags:
      - tables
      summary: Table details (screen 07)
      description: |-
        Same shape as a grid card. For the lines, call `GET /api/orders/{current_order.id}`. An inactive table is `NOT_FOUND` for a CASHIER (BR-TBL-06).

        BR: BR-TBL-01, BR-TBL-04, BR-TBL-06.
        Errors: `UNAUTHENTICATED`, `NOT_FOUND`.
      x-roles:
      - ADMIN
      - CASHIER
      x-br:
      - BR-TBL-01
      - BR-TBL-04
      - BR-TBL-06
      x-error-codes:
      - UNAUTHENTICATED
      - NOT_FOUND
      responses:
        '200':
          description: The table.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
    patch:
      operationId: updateTable
      tags:
      - tables
      summary: Renumber / activate / deactivate a table
      description: |-
        An OCCUPIED table cannot be deactivated or renumbered → `TABLE_OCCUPIED` (BR-TBL-06). Number unique (`DUPLICATE_VALUE`). No delete exists (OQ-32, BR-GEN-04). ADMIN only.

        BR: BR-ROLE-01, BR-TBL-06, BR-GEN-04, OQ-32.
        Errors: `UNAUTHENTICATED`, `FORBIDDEN_ROLE`, `NOT_FOUND`, `TABLE_OCCUPIED`, `DUPLICATE_VALUE`, `VALIDATION_ERROR`.
      x-roles:
      - ADMIN
      x-br:
      - BR-ROLE-01
      - BR-TBL-06
      - BR-GEN-04
      - OQ-32
      x-error-codes:
      - UNAUTHENTICATED
      - FORBIDDEN_ROLE
      - NOT_FOUND
      - TABLE_OCCUPIED
      - DUPLICATE_VALUE
      - VALIDATION_ERROR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableUpdate'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableEnvelope'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: pos_refresh
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
    Error401:
      description: INVALID_CREDENTIALS (login) or UNAUTHENTICATED (missing/invalid/expired/revoked session).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error403:
      description: FORBIDDEN_ROLE, NOT_ORDER_OWNER, or USER_DISABLED (login).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error404:
      description: NOT_FOUND.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error409:
      description: State conflict — see the operation's x-error-codes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Error422:
      description: VALIDATION_ERROR, ORDER_EMPTY, DISCOUNT_INVALID or INSUFFICIENT_CASH.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    TableStatus:
      type: string
      enum:
      - AVAILABLE
      - OCCUPIED
      description: Derived, never stored (BR-TBL-01).
...
    ErrorCode:
      type: string
      description: BR §9 codes (BR-GEN-06). The HTTP status of each code is fixed by BR §9. `INTERNAL_ERROR` (500) is technical,
        not a business rejection; approved by the owner at GATE B (D3) and added to BR §9 in BR v1.3 (ADR-0010).
      enum:
      - INVALID_CREDENTIALS
      - USER_DISABLED
      - TOO_MANY_ATTEMPTS
      - FORBIDDEN_ROLE
      - SHIFT_ALREADY_OPEN
      - NO_OPEN_SHIFT
      - SHIFT_HAS_OPEN_ORDERS
      - TABLE_OCCUPIED
      - TABLE_INACTIVE
      - ITEM_INACTIVE
      - ORDER_EMPTY
      - ORDER_NOT_EDITABLE
      - NOT_ORDER_OWNER
      - DISCOUNT_INVALID
      - INSUFFICIENT_CASH
      - ITEM_IN_USE
      - CATEGORY_NOT_EMPTY
      - USER_HAS_OPEN_SHIFT
      - LAST_ADMIN
      - VALIDATION_ERROR
      - UNAUTHENTICATED
      - NOT_FOUND
      - DUPLICATE_VALUE
      - IDEMPOTENCY_CONFLICT
      - INTERNAL_ERROR
...
    ErrorDetails:
      type: object
      description: Always present, may be empty. Known keys per code (ADR-0010).
      properties:
        fields:
          type: object
          description: 'VALIDATION_ERROR: field path → machine hints (English, never shown).'
          additionalProperties:
            type: array
            items:
              type: string
        open_order_numbers:
          type: array
          items:
            type: integer
          description: SHIFT_HAS_OPEN_ORDERS (BR-SHIFT-04).
        retry_after_seconds:
          type: integer
          nullable: true
          description: TOO_MANY_ATTEMPTS.
        item_ids:
          type: array
          items:
            type: integer
          description: ITEM_INACTIVE (BR-ITEM-01).
        field:
          type: string
          description: 'DUPLICATE_VALUE: the unique field (BR §9).'
      additionalProperties: true
    Error:
      type: object
      required:
      - success
      - code
      - message
      - details
      properties:
        success:
          type: boolean
          enum:
          - false
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Arabic message from BR §9, verbatim.
        details:
          $ref: '#/components/schemas/ErrorDetails'
...
    TableCurrentOrder:
      type: object
      required:
      - id
      - number
      - item_count
      - total_minor
      - cashier_id
      - cashier_name
      - created_at
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        number:
          type: integer
          description: Order number, e.g. 1048 (BR-ORD-10).
        item_count:
          type: integer
          description: Σ qty (BR-TBL-04).
        total_minor:
          type: integer
          format: int64
          description: Order total after discount (BR-TBL-04, OQ-14). Integer minor units (BR-GEN-01).
        cashier_id:
          type: integer
          format: int64
          minimum: 1
        cashier_name:
          type: string
        created_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
    Table:
      type: object
      required:
      - id
      - number
      - is_active
      - status
      - current_order
      - created_at
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        number:
          type: integer
          minimum: 1
        is_active:
          type: boolean
        status:
          $ref: '#/components/schemas/TableStatus'
        current_order:
          allOf:
          - $ref: '#/components/schemas/TableCurrentOrder'
          nullable: true
        created_at:
          type: string
          format: date-time
          description: RFC 3339 UTC (BR-GEN-02).
    TableCounts:
      type: object
      required:
      - all
      - available
      - occupied
      properties:
        all:
          type: integer
        available:
          type: integer
        occupied:
          type: integer
    TableList:
      type: object
      description: Not paginated (a floor has few tables). Ordered by number ascending.
      required:
      - items
      - counts
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Table'
        counts:
          $ref: '#/components/schemas/TableCounts'
    TableCreate:
      type: object
      required:
      - number
      properties:
        number:
          type: integer
          minimum: 1
        is_active:
          type: boolean
          default: true
    TableUpdate:
      type: object
      properties:
        number:
          type: integer
          minimum: 1
        is_active:
          type: boolean
...
    TableListEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/TableList'
    TableEnvelope:
      type: object
      required:
      - success
      - message
      - data
      properties:
        success:
          type: boolean
          enum:
          - true
        message:
          type: string
          description: Informational, never shown to users (ADR-0010).
        data:
          $ref: '#/components/schemas/Table'
...